+++ iftttie/context.py
@@ -8,9 +8,9 @@ from typing import Any, Awaitable, Callable, Iterable, Optional, Sequence, Tuple
 from aiohttp import ClientSession
 from loguru import logger
 
-from iftttie.database import insert_update
-from iftttie.services.base import BaseService
-from iftttie.types import Update
+from iftttie.channels.base import BaseChannel
+from iftttie.database import insert_event
+from iftttie.types import Event
 
 
 @dataclass
@@ -19,18 +19,18 @@ class Context:
     users: Sequence[Tuple[str, str]]
     db: Optional[Connection] = None
     session: Optional[ClientSession] = None
-    run_queue_task: Optional[Task] = None
-    services: Iterable[BaseService] = ()
-    on_update: Callable[[Update], Awaitable[Any]] = None
+    run_channels_task: Optional[Task] = None
+    channels: Iterable[BaseChannel] = ()
+    on_event: Callable[[Event], Awaitable[Any]] = None
 
-    async def on_event(self, update: Update):
+    async def trigger_event(self, event: Event):
         """
         Handle a single event in the application context.
         """
-        logger.success('{key} = {value!r}', key=update.key, value=update.value)
-        insert_update(self.db, update)
-        if self.on_update is not None:
+        logger.success('{key} = {value!r}', key=event.key, value=event.value)
+        insert_event(self.db, event)
+        if self.on_event is not None:
             try:
-                await self.on_update(update)
+                await self.on_event(event)
             except Exception as e:
                 logger.opt(exception=e).error('Error while handling the event.')